Sort eslint errors
